Adding output options

+++ fetch_authorities.py
@@ -11,6 +11,9 @@ vocabularies = parser.add_mutually_exclusive_group()
 vocabularies.add_argument('--lcnaf-only', help='only download name authority records', action='store_true')
 vocabularies.add_argument('--lcsh-only', help='only download subject authority records', action='store_true')
 vocabularies.add_argument('--lcgft-only', help='only download genre/form authority records', action='store_true')
+print_opts = parser.add_argument_group(title='Output options')
+print_opts.add_argument('--print-matched-headings', '-m', help='Print headings that are matched and downloaded to STDOUT', action='store_true')
+print_opts.add_argument('--print-unmatched-headings', '-u', help='Print headings that have no LC matches to STDOUT', action='store_true')
 args = parser.parse_args()
 
 try:
@@ -26,28 +29,32 @@ try:
             heading_suffix = ''
 
         for term in authorized_terms:
-            print(term)
             response = search_client.search(term.lstrip().rstrip('.,;') + heading_suffix)
-            i = 0
-            for result in response:
-                if i < args.matches:
-                    i = i + 1
-                    if 'authorities' in result.link:
-                        auth_id = re.search('.*\/authorities\/.*\/(.*)', result.link).group(1)
-                        auth_file_url = 'http://lccn.loc.gov/' + auth_id + '/marcxml'
-                        time.sleep(10)
-                        urllib.urlretrieve (auth_file_url, 'tmp.marcxml')
-                        try:
-                            records = marcxml.parse_xml_to_array('tmp.marcxml')
-                            writer = MARCWriter(file('authorities_to_load.mrc', 'ab'))
-                            for r in records:
-                                writer.write(r)
-                                writer.close()
-                            os.remove('tmp.marcxml')
-                        except SAXParseException:
-                            print('Could not parse MARCXML file')
-                else:
-                    break
+            if response.totalResults:
+                if args.print_matched_headings:
+                    print(term)
+                i = 0
+                for result in response:
+                    if i < args.matches:
+                        i = i + 1
+                        if 'authorities' in result.link:
+                            auth_id = re.search('.*\/authorities\/.*\/(.*)', result.link).group(1)
+                            auth_file_url = 'http://lccn.loc.gov/' + auth_id + '/marcxml'
+                            time.sleep(10)
+                            urllib.urlretrieve (auth_file_url, 'tmp.marcxml')
+                            try:
+                                records = marcxml.parse_xml_to_array('tmp.marcxml')
+                                writer = MARCWriter(file('authorities_to_load.mrc', 'ab'))
+                                for r in records:
+                                    writer.write(r)
+                                    writer.close()
+                                os.remove('tmp.marcxml')
+                            except SAXParseException:
+                                print('Could not parse MARCXML file')
+                    else:
+                        break
+            elif args.print_unmatched_headings:
+                print(term)
 except IOError:
     print("File '%s' doesn't exist", args.filename)
     exit(1)
